Add iSim / physical sim selection.
Modified to have the SIM selection as a compile time option via the mbed_lib.json file

Set the sim_select variable to wither 0 (for external) of 5 (for iSim).

The setting is made during initialisation, but the Modem needs to be reset for it to take effect.
For now the device needs to be power cycled for this.